添加龙虎榜

+++ stanalyse/src/main/java/com/wy/stock/tools/StockDownloadToolTHSImpl.java
@@ -518,10 +518,9 @@ public class StockDownloadToolTHSImpl implements StockDownloadToolTHS {
 		url = "http://data.10jqka.com.cn/ifmarket/lhbggxq/report/" + dateStr;
 		file = new File(dirPath + StockConstant.LHB_FILE_NAME);
 		try {
-			if(StockUtils.getTotalLines(dirPath + StockConstant.LHB_FILE_NAME) < 2000){
+			if(file.exists() && StockUtils.getTotalLines(dirPath + StockConstant.LHB_FILE_NAME) < 2000){
 				file.delete();
 			}
-			// 已经存在的不再重新下载
 			if(!StringUtils.isEmpty(url) && file != null && !file.exists()){
 				HttpUtils.httpDownload(url, "GB2312", 10 * 1000, file);
 			}
